add nft to extractor
move Extractor function to db
