+++ db/extract.go
@@ -1,8 +1,98 @@
 package db
 
 import (
+	"encoding/json"
+	"fmt"
+	"time"
+
+	"github.com/cosmos/cosmos-sdk/types"
+	"github.com/jackc/pgtype"
 	"github.com/jackc/pgx/v4/pgxpool"
+	"github.com/likecoin/likecoin-chain-tx-indexer/logger"
+	"github.com/likecoin/likecoin-chain-tx-indexer/utils"
 )
+
+// todo: move to config
+const LIMIT = 10000
+
+type EventHandler func(batch *Batch, message []byte, events types.StringEvents, timestamp time.Time) error
+
+func Extract(conn *pgxpool.Conn, handlers map[string]EventHandler) (finished bool, err error) {
+	begin, err := GetMetaHeight(conn, "iscn")
+	if err != nil {
+		return false, fmt.Errorf("Failed to get ISCN synchonized height: %w", err)
+	}
+
+	end, err := GetLatestHeight(conn)
+	if err != nil {
+		return false, fmt.Errorf("Failed to get latest height: %w", err)
+	}
+	if begin == end {
+		return true, nil
+	}
+	if begin+LIMIT < end {
+		end = begin + LIMIT
+	} else {
+		finished = true
+	}
+
+	ctx, _ := GetTimeoutContext()
+
+	sql := `
+	SELECT height, tx #> '{"tx", "body", "messages"}' AS messages, tx -> 'logs' AS logs, tx #> '{"timestamp"}' as timestamp
+	FROM txs
+	WHERE height >= $1
+		AND height < $2
+	ORDER BY height ASC;
+	`
+
+	rows, err := conn.Query(ctx, sql, begin, end)
+	defer rows.Close()
+
+	batch := NewBatch(conn, LIMIT)
+
+	for rows.Next() {
+		var height int64
+		var messageData pgtype.JSONB
+		var eventData pgtype.JSONB
+		var timestamp time.Time
+		err := rows.Scan(&height, &messageData, &eventData, &timestamp)
+		if err != nil {
+			panic(err)
+		}
+
+		var messages []json.RawMessage
+		err = messageData.AssignTo(&messages)
+		if err != nil {
+			panic(err)
+		}
+		var events []struct {
+			Events types.StringEvents `json:"events"`
+		}
+
+		err = eventData.AssignTo(&events)
+		if err != nil {
+			panic(err)
+		}
+
+		for i, event := range events {
+			action := utils.GetEventsValue(event.Events, "message", "action")
+			if handler, ok := handlers[action]; ok {
+				err = handler(&batch, messages[i], event.Events, timestamp)
+				if err != nil {
+					panic(err)
+				}
+			}
+		}
+	}
+	batch.UpdateISCNHeight(end)
+	logger.L.Infof("ISCN synced height: %d", end)
+	err = batch.Flush()
+	if err != nil {
+		return false, fmt.Errorf("send batch failed: %w", err)
+	}
+	return finished, nil
+}
 
 func GetMetaHeight(conn *pgxpool.Conn, key string) (int64, error) {
 	ctx, _ := GetTimeoutContext()
